create a Base class to inherintance
