+++ lib/zx/core.rb
@@ -2,6 +2,8 @@
 
 module Core
   require_relative 'core/util/parameter'
+
+  require_relative 'core/base'
   require_relative 'core/and_then'
   require_relative 'core/caller'
   require_relative 'core/match'
